+++ packages/static/package.json
@@ -26,7 +26,7 @@
     "./dist"
   ],
   "scripts": {
-    "build": "tsc && vite build",
+    "build": "pnpm run generate-theme && tsc && vite build",
     "build-storybook": "storybook build",
     "dev:vite": "vite",
     "generate-theme": "node src/scripts/generate-theme.js",
