+++ packages/static/src/chip/styles/chip.module.css
@@ -1,4 +1,4 @@
-@import '../../theme-static/theme.module.css';
+@import '../../theme/theme.module.css';
 
 .chip {
   border: 1px solid rgb(230, 230, 230);
