+++ packages/static/src/index.ts
@@ -2,4 +2,4 @@ export {default as Chip} from './chip'
 export type {ChipProps} from './chip'
 export {default as Popover} from './popover'
 export type {PopoverDirection, PopoverProps} from './popover'
-export {default as ThemeStaticProvider} from './theme-static/ThemeProvider'
+export {default as ThemeStaticProvider} from './theme/ThemeProvider'
