+++ packages/static/src/theme/theme.module.css
@@ -1,4 +1,4 @@
-// THIS FILE IS AUTO GENERATED. DO NOT MODIFY!
+/* THIS FILE IS AUTO GENERATED. DO NOT MODIFY!  */
 :root {
   /* Spacings */
   --cdg-spacing-1 : 0.25rem;
