+++ packages/static/src/utils/objectToCss/object-to-css.ts
@@ -1,4 +1,4 @@
-import theme from '../../theme-static/theme'
+import theme from '../../theme/theme'
 export interface StyleObject {
   [key: string]: string | StyleObject
 }
